Move ExtraTypeInfo to separate file
