+++ src/include/duckdb/common/extra_type_info.hpp
@@ -1,0 +1,197 @@
+//===----------------------------------------------------------------------===//
+//                         DuckDB
+//
+// duckdb/common/extra_type_info.hpp
+//
+//
+//===----------------------------------------------------------------------===//
+
+#pragma once
+
+#include "duckdb/common/common.hpp"
+
+namespace duckdb {
+
+//! Extra Type Info Type
+enum class ExtraTypeInfoType : uint8_t {
+	INVALID_TYPE_INFO = 0,
+	GENERIC_TYPE_INFO = 1,
+	DECIMAL_TYPE_INFO = 2,
+	STRING_TYPE_INFO = 3,
+	LIST_TYPE_INFO = 4,
+	STRUCT_TYPE_INFO = 5,
+	ENUM_TYPE_INFO = 6,
+	USER_TYPE_INFO = 7,
+	AGGREGATE_STATE_TYPE_INFO = 8
+};
+
+
+struct ExtraTypeInfo {
+	explicit ExtraTypeInfo(ExtraTypeInfoType type);
+	explicit ExtraTypeInfo(ExtraTypeInfoType type, string alias);
+	virtual ~ExtraTypeInfo();
+
+	ExtraTypeInfoType type;
+	string alias;
+	optional_ptr<TypeCatalogEntry> catalog_entry;
+
+public:
+	bool Equals(ExtraTypeInfo *other_p) const;
+	//! Serializes a ExtraTypeInfo to a stand-alone binary blob
+	virtual void Serialize(FieldWriter &writer) const;
+	//! Serializes a ExtraTypeInfo to a stand-alone binary blob
+	static void Serialize(ExtraTypeInfo *info, FieldWriter &writer);
+	//! Deserializes a blob back into an ExtraTypeInfo
+	static shared_ptr<ExtraTypeInfo> Deserialize(FieldReader &reader);
+
+	virtual void FormatSerialize(FormatSerializer &serializer) const;
+	static shared_ptr<ExtraTypeInfo> FormatDeserialize(FormatDeserializer &source);
+
+	template <class TARGET>
+	TARGET &Cast() {
+		D_ASSERT(dynamic_cast<TARGET *>(this));
+		return reinterpret_cast<TARGET &>(*this);
+	}
+	template <class TARGET>
+	const TARGET &Cast() const {
+		D_ASSERT(dynamic_cast<const TARGET *>(this));
+		return reinterpret_cast<const TARGET &>(*this);
+	}
+
+protected:
+	virtual bool EqualsInternal(ExtraTypeInfo *other_p) const;
+};
+
+struct DecimalTypeInfo : public ExtraTypeInfo {
+	DecimalTypeInfo(uint8_t width_p, uint8_t scale_p);
+
+	uint8_t width;
+	uint8_t scale;
+
+public:
+	void Serialize(FieldWriter &writer) const override;
+	void FormatSerialize(FormatSerializer &serializer) const override;
+
+	static shared_ptr<ExtraTypeInfo> FormatDeserialize(FormatDeserializer &source);
+	static shared_ptr<ExtraTypeInfo> Deserialize(FieldReader &reader);
+
+protected:
+	bool EqualsInternal(ExtraTypeInfo *other_p) const override;
+};
+
+struct StringTypeInfo : public ExtraTypeInfo {
+	explicit StringTypeInfo(string collation_p);
+
+	string collation;
+
+public:
+	void Serialize(FieldWriter &writer) const override;
+	static shared_ptr<ExtraTypeInfo> Deserialize(FieldReader &reader);
+
+	void FormatSerialize(FormatSerializer &serializer) const override;
+	static shared_ptr<ExtraTypeInfo> FormatDeserialize(FormatDeserializer &source);
+
+protected:
+	bool EqualsInternal(ExtraTypeInfo *other_p) const override;
+};
+
+struct ListTypeInfo : public ExtraTypeInfo {
+	explicit ListTypeInfo(LogicalType child_type_p);
+
+	LogicalType child_type;
+
+public:
+	void Serialize(FieldWriter &writer) const override;
+	void FormatSerialize(FormatSerializer &serializer) const override ;
+
+	static shared_ptr<ExtraTypeInfo> Deserialize(FieldReader &reader);
+	static shared_ptr<ExtraTypeInfo> FormatDeserialize(FormatDeserializer &source);
+
+protected:
+	bool EqualsInternal(ExtraTypeInfo *other_p) const override;
+};
+
+struct StructTypeInfo : public ExtraTypeInfo {
+	explicit StructTypeInfo(child_list_t<LogicalType> child_types_p);
+
+	child_list_t<LogicalType> child_types;
+
+public:
+	void Serialize(FieldWriter &writer) const override;
+	void FormatSerialize(FormatSerializer &serializer) const override;
+
+	static shared_ptr<ExtraTypeInfo> Deserialize(FieldReader &reader);
+
+	static shared_ptr<ExtraTypeInfo> FormatDeserialize(FormatDeserializer &deserializer);
+
+protected:
+	bool EqualsInternal(ExtraTypeInfo *other_p) const override;
+};
+
+struct AggregateStateTypeInfo : public ExtraTypeInfo {
+	explicit AggregateStateTypeInfo(aggregate_state_t state_type_p);
+
+	aggregate_state_t state_type;
+
+public:
+	void Serialize(FieldWriter &writer) const override;
+	void FormatSerialize(FormatSerializer &serializer) const override ;
+
+	static shared_ptr<ExtraTypeInfo> FormatDeserialize(FormatDeserializer &source);
+	static shared_ptr<ExtraTypeInfo> Deserialize(FieldReader &reader);
+
+protected:
+	bool EqualsInternal(ExtraTypeInfo *other_p) const override;
+};
+
+struct UserTypeInfo : public ExtraTypeInfo {
+	explicit UserTypeInfo(string name_p);
+
+	string user_type_name;
+
+public:
+	void Serialize(FieldWriter &writer) const override;
+	void FormatSerialize(FormatSerializer &serializer) const override;
+
+	static shared_ptr<ExtraTypeInfo> Deserialize(FieldReader &reader);
+	static shared_ptr<ExtraTypeInfo> FormatDeserialize(FormatDeserializer &source);
+
+protected:
+	bool EqualsInternal(ExtraTypeInfo *other_p) const override;
+};
+
+// If this type is primarily stored in the catalog or not. Enums from Pandas/Factors are not in the catalog.
+enum EnumDictType : uint8_t { INVALID = 0, VECTOR_DICT = 1 };
+
+struct EnumTypeInfo : public ExtraTypeInfo {
+	explicit EnumTypeInfo(string enum_name_p, Vector &values_insert_order_p, idx_t dict_size_p);
+	EnumTypeInfo(const EnumTypeInfo &) = delete;
+	EnumTypeInfo &operator=(const EnumTypeInfo &) = delete;
+
+public:
+	const EnumDictType &GetEnumDictType() const;
+	const string &GetEnumName() const;
+	const string GetSchemaName() const;
+	const Vector &GetValuesInsertOrder() const;
+	const idx_t &GetDictSize() const;
+
+	static LogicalType CreateType(const string &enum_name, Vector &ordered_data, idx_t size);
+
+	void Serialize(FieldWriter &writer) const override;
+	static shared_ptr<ExtraTypeInfo> Deserialize(FieldReader &reader);
+
+	void FormatSerialize(FormatSerializer &serializer) const override;
+	static shared_ptr<ExtraTypeInfo> FormatDeserialize(FormatDeserializer &source);
+
+protected:
+	// Equalities are only used in enums with different catalog entries
+	bool EqualsInternal(ExtraTypeInfo *other_p) const override;
+
+	Vector values_insert_order;
+private:
+	EnumDictType dict_type;
+	string enum_name;
+	idx_t dict_size;
+};
+
+} // namespace duckdb
